+++ src/weightlens/sources/pytorch.py
@@ -53,7 +53,7 @@ class PyTorchWeightSource(WeightSource):
                     continue
 
                 materialized = tensor.detach().cpu().contiguous()
-                values = materialized.reshape(-1).numpy()
+                values = materialized.reshape(-1).float().numpy()
                 layer_count += 1
                 logger.debug(
                     "Yielding layer %s shape=%s dtype=%s param_count=%d.",
